Add user supplied documentation of WinSCP
The Legacy after Nansen found the documentation of WinSCP on the Sigma2
pages meager. They suggested to write documentation they found
necessary, RT-case 185392.

Their contribution (https://github.com/SIOS-Svalbard/NIRD_doc.git) is
added as a submodule to the repository.

The original nird file is split over several files starting with a
number 01 to 03. The 05_winscp.md is a link to the file in the SIOS
repository. All links are in the 06_links.md file, which is read last by
pandoc. The generation of the new nird.md file is done with
pandoc 0[1-6]*.md -o nird.md

The original nird.md file is move to nird_not_generated.