move APR documentation from code_development/ to jobs/
I think Arm Performance Reports are primarily for users, not only for
developers. It is also useful for developers but I want this to be more
visible and not "hidden" at the end of the documentation under "CODE
DEVELOPMENT".

This tool will also be used for large allocation applications 2020.2 and
therefore needs to be more visible.

Also clarified and updated documentation for Fram (where this seems to
work in compatibility mode) and Saga (where this is not working since we
cannot use the compatibility mode).