rename section
To a hopefully better title since I would like users
to read this when they are about to select job parameters,
not only as hindsight.

I haven't changed the text yet but I will try to make this
page about how to choose the right amount of memory and how
to find out how much memory a job needs and why it matters.

We then need a similar page of selecting the number of cores
and some information is already here and there and I will try to collect
that in a second step.

Although the memory and core settings are not completely independent,
I will try to discuss them as independently as possible so that both
of these can be consumed and examined independently.