Added guide to explain how to use TensorFlow
The guide explains how to deploy TensorFlow on Saga, preferably using
GPU resources. The guide goes through how to load/install TensorFlow,
how to load datasets into TensorFlow, how to save and checkpoint data
and how to distribute the computation over multiple GPUs.

One thing that is missing from the current guide is how to distribute
over more than four GPUs. This was left out as the current TensorFlow
support is lacking and we are exploring other libraries for enabling
this.