Added section on Mosh
Mosh is a remote terminal application that supports intermittent
connections and is perfect for connecting to Sigma 2 resources from a
laptop. Making users aware of this option is important.

The changes affected two pages, one with the information and one with a
link to the new information. Additional formatting of these two files
were also performed.